requirements.txt
streamlit
pandas
numpy
random
matplotlib.pyplot
statistics
